feat(api): Introduce networking API
This commit introduces the definitions for the underlying system that
will represent proper networking support in KraftKit.  These definitions
allow for defining both networks and interfaces to a network which are
manipulated through the Zip API Framework[0].

Simultaneously, the list of network interfaces is added to the
MachineSpec which will be used to iterate against networks during the
instance's instantiation.

[0]: https://api.zip

Signed-off-by: Alexander Jung <alex@unikraft.io>