feat(qemu): Introduce volume support
This commit adjusts the current QEMU driver implementation for the
machine interface to recognise the supplied volume templates within the
machine specification.

Currently, only the 9pfs driver is handled but the implementation is
made to easily allow for more.  To make this integration possible,
correctly handling the filesystem devices via the `-fsdev` flag is
required and so necessary structures are introduced to handle this in
the same idiomatic manner found in the rest of the QEMU driver
implementation.

Signed-off-by: Alexander Jung <alex@unikraft.io>