feat!: Introduce dynamic initramfs serialization for root filesystems
This commit introduces a refactor of the mechanics behind serializing an
input "rootfs" such that it is a usable initramfs.  This change starts
with the adjustment of the `initrd` package which is now based on an
interface `Initrd` which is used to introduce several underlying
implementations that can serialize a provided input.  The following
possible inputs are:

- A path to a directory which will act as the rootfs;
- A `Dockerfile` which can be built via buildkit; and,
- For compatibility of the interface, an existing file representing a
  CPIO archive.

The adjustment to this interface has been adjusted across the project
including relevant tools.

Signed-off-by: Alexander Jung <alex@unikraft.io>