fix(actions): Fix accessing access to credentials
In a typical usecase and to make KraftKit's GitHub action a drop-in
replacement with existing pipelines, the use of the Docker Login[0]
action is typically used which stores credentials in the well-known
path `$HOME/.docker/config.json`.  However, since the action is built
as a container image, this path is not accessible during the container's
instantiation.  To make accessing this configuration possible, turn
the action into a composite with a script that is executed before the
container image which copies these credentials into the pwth that
represents the new `$HOME` of the container image.

[0]: https://github.com/docker/login-action

Signed-off-by: Alexander Jung <alex@unikraft.io>