refactor(build): Abstract `kraft build` into interface implementations
This commit refactors the subcommand `kraft build` into an interface
`builder` which is developed in a similar vein of the interface to
`kraft run`'s `runner` such that different contexts can be used to
determine whether the provided working directory or positional
arguments are "buildable".

In this first iteration, the existing context of a working directory
populated with a `Kraftfile` that contains a `unikraft` element is
abstracted.  This is the current mode for building a Unikraft project,
however future contexts exist as well which can be developed, including:
`Kraftfile`'s that only supply a `template` element; `Kraftfile`'s that
only supply a `runtime` element (and do not construct a unikernel but do
construct a root filesystem); or the absense of a `Kraftfile` entirely
and intelligently determining the relevant template based on working
directory source files (such as determining a Go-based by project by
seeking a top-level `go.mod`).

Signed-off-by: Alexander Jung <alex@unikraft.io>